// backend/src/types/amadeus.ts
export interface HotelSearchParams {
  cityCode: string;
  checkInDate: string;
  checkOutDate: string;
  adults: number;
  children?: number;
  radius?: number;
  radiusUnit?: 'KM' | 'MILE';
  hotelSource?: string;
  ratings?: number[];
  priceRange?: string;
  currency?: string;
  paymentPolicy?: string;
  boardType?: string;
}

export interface Hotel {
  type: string;
  hotelId: string;
  chainCode: string;
  dupeId: string;
  name: string;
  rating: string;
  cityCode: string;
  latitude: number;
  longitude: number;
  hotelDistance: HotelDistance;
  address: HotelAddress;
  contact: HotelContact;
  description: HotelDescription;
  amenities: string[];
  media: HotelMedia[];
}

export interface HotelDistance {
  value: number;
  unit: string;
}

export interface HotelAddress {
  lines: string[];
  postalCode: string;
  cityName: string;
  countryCode: string;
}

export interface HotelContact {
  phone: string;
  fax: string;
  email: string;
}

export interface HotelDescription {
  lang: string;
  text: string;
}

export interface HotelMedia {
  uri: string;
  category: string;
}